move a class from ssh-slaves
